pyproject.toml
[tool.poetry]
name = "projeto-01-exemplo"
version = "0.1.0"
description = ""
authors = ["luciano"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
apache-airflow = "^2.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
